refactor: ServiceRequestDto 위치 변경
