+++ src/main/java/com/example/extra/domain/schedule/dto/service/request/ScheduleCreateServiceRequestDto.java
@@ -1,4 +1,4 @@
-package com.example.extra.domain.schedule.dto.service;
+package com.example.extra.domain.schedule.dto.service.request;
 
 import java.time.LocalDate;
