+++ src/main/java/com/example/extra/domain/schedule/service/impl/ScheduleServiceImpl.java
@@ -6,7 +6,7 @@ import com.example.extra.domain.jobpost.entity.JobPost;
 import com.example.extra.domain.jobpost.exception.JobPostErrorCode;
 import com.example.extra.domain.jobpost.exception.NotFoundJobPostException;
 import com.example.extra.domain.jobpost.repository.JobPostRepository;
-import com.example.extra.domain.schedule.dto.service.ScheduleCreateServiceRequestDto;
+import com.example.extra.domain.schedule.dto.service.request.ScheduleCreateServiceRequestDto;
 import com.example.extra.domain.schedule.entity.Schedule;
 import com.example.extra.domain.schedule.mapper.service.ScheduleEntityMapper;
 import com.example.extra.domain.schedule.repository.ScheduleRepository;
